8343779: Cannot build Vector API JMH benchmarks on vectorIntrinsics branch

+++ test/micro/org/openjdk/bench/jdk/incubator/vector/bigdata/VectorDistance.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -65,11 +65,11 @@ public class VectorDistance {
             int l;
             l = input[i << 2];
             l &= 0xff;
-            l |= ((long) input[(i << 2) + 1] << 8);
+            l |= ((int) input[(i << 2) + 1] << 8);
             l &= 0xffff;
-            l |= ((long) input[(i << 2) + 2] << 16);
+            l |= ((int) input[(i << 2) + 2] << 16);
             l &= 0xffffff;
-            l |= ((long) input[(i << 2) + 3] << 24);
+            l |= ((int) input[(i << 2) + 3] << 24);
             floatArr[i] = Float.intBitsToFloat(l);
         }
         return floatArr;
